Edit read user details and delete user on the web (address -> addressSubdistrictId, yearsOfExperience -> yearsOfExperiences, expectedSalaryInMillionVnd -> expectedSalaryInVnd)

+++ routes/users.js
@@ -60,7 +60,10 @@ router.get('/', function(req, res, next) {
 
         "where " +
             commonResources.USERS_COLUMN_FULL_NAME + " " +
-                "like '%" + keyword + "%';";
+                "like '%" + keyword + "%' " +
+        "order by " +
+            commonResources.USERS_TABLE_NAME + "." +
+            commonResources.USERS_COLUMN_ID + " desc";
 
     dbConnect.query(getAllUsersSql, function (err, result, fields) {
        if (err) throw err;
@@ -82,7 +85,7 @@ router.post('/details', async (req, res) => {
                 + commonResources.GENDERS_COLUMN_NAME + " as " +
                 commonResources.COLUMN_ALIAS_GENDER + ", " +
 
-            commonResources.USERS_COLUMN_EXPECTED_SALARY_MIL_VND + ", " +
+            commonResources.USERS_COLUMN_EXPECTED_SALARY_VND + ", " +
 
             // TypesOfWork.name as typeOfWork
             commonResources.TYPES_OF_WORK_TABLE_NAME + "." +
@@ -92,10 +95,22 @@ router.post('/details', async (req, res) => {
             commonResources.LEVELS_OF_EDUCATION_COLUMN_NAME + " as " +
                 commonResources.COLUMN_ALIAS_GRADUATED_EDUCATION_LEVEL + ", " +
 
-            commonResources.USERS_COLUMN_YEARS_OF_EXPERIENCE + ", " +
+            commonResources.USERS_COLUMN_YEARS_OF_EXPERIENCES + ", " +
             commonResources.USERS_COLUMN_RESUME_SUMMARY + ", " +
             commonResources.USERS_COLUMN_DOB_MILLIS + ", " +
-            commonResources.USERS_COLUMN_ADDRESS + ", " +
+
+           commonResources.STATE_PROVINCES_TABLE_NAME + "." +
+           commonResources.STATE_PROVINCES_COLUMN_NAME + " as " +
+           commonResources.COLUMN_ALIAS_STATE_PROVINCE_NAME + ", " +
+
+           commonResources.DISTRICTS_TABLE_NAME + "." +
+           commonResources.DISTRICTS_COLUMN_NAME + " as " +
+           commonResources.COLUMN_ALIAS_DISTRICT_NAME + ", " +
+
+           commonResources.SUBDISTRICTS_TABLE_NAME + "." +
+           commonResources.SUBDISTRICTS_COLUMN_NAME + " as " +
+           commonResources.COLUMN_ALIAS_SUBDISTRICT_NAME + ", " +
+
             commonResources.USERS_COLUMN_PHONE + ", " +
             commonResources.USERS_COLUMN_EMAIL + ", " +
             commonResources.USERS_COLUMN_AVATAR_URL + ", " +
@@ -117,6 +132,21 @@ router.post('/details', async (req, res) => {
                 commonResources.GENDERS_TABLE_NAME + "." +
                 commonResources.GENDERS_COLUMN_ID + " = " +
                 commonResources.USERS_COLUMN_GENDER_ID + " " +
+           "left join " +
+                commonResources.SUBDISTRICTS_TABLE_NAME + " " + "on " +
+               commonResources.USERS_COLUMN_ADDRESS_SUBDISTRICT_ID +
+               " = " + commonResources.SUBDISTRICTS_TABLE_NAME + "." +
+               commonResources.SUBDISTRICTS_COLUMN_ID + " " +
+           "left join " + commonResources.DISTRICTS_TABLE_NAME +
+               " on " + commonResources.SUBDISTRICTS_TABLE_NAME + "." +
+               commonResources.SUBDISTRICTS_COLUMN_DISTRICT_ID + " = " +
+               commonResources.DISTRICTS_TABLE_NAME + "." +
+               commonResources.DISTRICTS_COLUMN_ID + " " +
+           "left join " + commonResources.STATE_PROVINCES_TABLE_NAME +
+               " on " + commonResources.DISTRICTS_TABLE_NAME + "." +
+               commonResources.DISTRICTS_COLUMN_STATE_PROVINCE_ID +
+               " = " + commonResources.STATE_PROVINCES_TABLE_NAME +
+               "." + commonResources.STATE_PROVINCES_COLUMN_ID + " " +
        "where Users.id = ?;";
 
    dbConnect.query(
@@ -125,14 +155,9 @@ router.post('/details', async (req, res) => {
        function (err, userResult) {
            if (err) throw err;
            let user = userResult[0]; // result is an array
-           if (user.birthdayInMilliseconds) {
-               user.birthdayInMilliseconds =
-                               moment(user.birthdayInMilliseconds)
-                               .format('DD-MMM-YYYY');
-           }
 
            // Experiences
-           let selectExperiencesByUserIdSql = 
+           let selectExperiencesByUserIdSql =
                "select " +
                     commonResources.EXPERIENCES_COLUMN_COMPANY_NAME + ", " +
                     commonResources.EXPERIENCES_COLUMN_JOB_TITLE + ", " +
@@ -145,6 +170,7 @@ router.post('/details', async (req, res) => {
                "order by "
                     + commonResources.EXPERIENCES_COLUMN_DATE_IN_MILLIS + " desc;";
 
+
            dbConnect.query(
                selectExperiencesByUserIdSql,
                [userId],
@@ -152,7 +178,7 @@ router.post('/details', async (req, res) => {
                    if (err) throw err;
                    let experiences = experiencesResult;
 
-                   let selectEducationByUserIdSql = 
+                   let selectEducationByUserIdSql =
                        "select " +
                             commonResources.EDUCATION_COLUMN_SCHOOL_NAME + ", " +
                             commonResources.EDUCATION_COLUMN_MAJOR + ", " +
@@ -195,20 +221,65 @@ router.post('/details', async (req, res) => {
                        [userId],
                        function (err, educationResult) {
                            if (err) throw err;
-                           res.render(
-                               'users/details',
-                               {
-                                        user, experiences,
-                                       educationResult, moment
-                                    }
-                               );
+
+                           // Query JobSkillsOfCandidate records
+                           // of this user
+                           let selectJobSkillsOfUserByUserIdSql =
+                               "select " +
+                                    commonResources.JOB_SKILLS_COLUMN_NAME + " " +
+                               "from " +
+                                    commonResources.JOB_SKILLS_OF_CANDIDATE_TABLE_NAME
+                                    + ", " +
+                                    commonResources.JOB_SKILLS_TABLE_NAME + " " +
+                               "where " +
+
+                                    commonResources.JOB_SKILLS_OF_CANDIDATE_TABLE_NAME
+                                    + "." +
+                                    commonResources
+                                        .JOB_SKILLS_OF_CANDIDATE_COLUMN_JOB_SKILLS_ID
+                                    + " = " +
+                                    commonResources.JOB_SKILLS_TABLE_NAME
+                                    + "."
+                                    + commonResources.JOB_SKILLS_COLUMN_ID
+                                    + " " +
+
+                                    "and " +
+                                    commonResources
+                                        .JOB_SKILLS_OF_CANDIDATE_COLUMN_USER_ID +
+                                    " = ?;";
+
+                           dbConnect.query(
+                               selectJobSkillsOfUserByUserIdSql,
+                               [userId],
+                               function (
+                                   selectJobSkillsErr,
+                                   selectJobSkillResult) {
+                                   if (selectJobSkillsErr) {
+                                       throw selectJobSkillsErr;
+                                   }
+                                   let jobSkills = selectJobSkillResult;
+                                   console.log({
+                                       user, experiences,
+                                       educationResult,
+                                       jobSkills
+                                   });
+                                   res.render(
+                                       'users/details',
+                                       {
+                                           user, experiences,
+                                           educationResult,
+                                           jobSkills,
+                                           moment
+                                       }
+                                   );
+                               }
+                           );
                        }
                    );
                }
            );
        }
    );
-
 });
 
 function checkIfPhoneExistsWhenCreateUser(phone, callback) {
@@ -334,15 +405,33 @@ router.post('/remove', async (req, res) => {
                         throw deleteExperienceErr;
                     }
 
-                    let deleteUserByIdSql =
-                        "delete from " + commonResources.USERS_TABLE_NAME + " " +
-                        "where " + commonResources.USERS_COLUMN_ID + " = ?;";
+                    let deleteJobSkillsByUserIdSql =
+                        "delete from " +
+                            commonResources
+                                .JOB_SKILLS_OF_CANDIDATE_TABLE_NAME + " " +
+                        "where " +
+                            commonResources
+                                .JOB_SKILLS_OF_CANDIDATE_COLUMN_USER_ID
+                            + " = ?;";
                     dbConnect.query(
-                        deleteUserByIdSql,
+                        deleteJobSkillsByUserIdSql,
                         [userId],
-                        function (err, result) {
-                            if (err) throw err;
-                            res.redirect('/users/');
+                        function (deleteJobSkillsErr, deleteJobSkillsResult) {
+                            if (deleteJobSkillsErr) {
+                                throw deleteJobSkillsErr;
+                            }
+
+                            let deleteUserByIdSql =
+                                "delete from " + commonResources.USERS_TABLE_NAME + " " +
+                                "where " + commonResources.USERS_COLUMN_ID + " = ?;";
+                            dbConnect.query(
+                                deleteUserByIdSql,
+                                [userId],
+                                function (err, result) {
+                                    if (err) throw err;
+                                    res.redirect('/users/');
+                                }
+                            );
                         }
                     );
                 }
